apps/agent-python/requirements.txt
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
cuid>=0.4.0
